+++ src/main/java/net/sodiumstudio/dwmg/registries/DwmgBefriendingTypes.java
@@ -142,5 +142,10 @@ public class DwmgBefriendingTypes {
 				ModEntityTypes.JACK_FROST.get(),
 				DwmgEntityTypes.HMAG_JACK_FROST.get(),
 				new HandlerJackFrost());
+		event.register(
+				ModEntityTypes.MELTY_MONSTER.get(),
+				DwmgEntityTypes.HMAG_MELTY_MONSTER.get(),
+				new HandlerJackFrost());// TODO: change
+		
 	}
 }